Moved years.js creation to main.py run
