+++ static/js/years.js
@@ -1,1 +1,1 @@
-var ACCYEARS = {"\u05e9\u05e0\u05ea 2013": ["01/01/2013", "31/12/2013"],"\u05e9\u05e0\u05ea 2012": ["01/01/2012", "31/12/2012"],"\u05e9\u05e0\u05ea 2011": ["01/01/2011", "31/12/2011"],"\u05e9\u05e0\u05ea 2010": ["01/01/2010", "31/12/2010"]};
+var ACCYEARS = {"\u05e9\u05e0\u05ea 2014": ["01/01/2014", "31/12/2014"], "\u05e9\u05e0\u05ea 2013": ["01/01/2013", "31/12/2013"], "\u05e9\u05e0\u05ea 2012": ["01/01/2012", "31/12/2012"], "\u05e9\u05e0\u05ea 2011": ["01/01/2011", "31/12/2011"]};
